mcp-servers/mem0-core/requirements.txt
mem0ai
mcp[cli]
